+++ Editor/UI/Scripts/Components/Tabs/SearchableTabs/HierarchyPopupTab.cs
@@ -1,11 +1,9 @@
-using System.Collections.Generic;
-using UnityEngine;
 
 namespace Unity.AI.Assistant.UI.Editor.Scripts.Components
 {
     class HierarchyPopupTab: SearchableTab
     {
-        public override string Filter => "h: ";
+        public override SearchContextWrapper[] SearchProviders => new[] { HierarchySearchContext };
         public override string DefaultLabel => "Hierarchy";
         public override int Order => 2;