+++ Editor/UI/Scripts/Components/Tabs/SearchableTabs/ProjectPopupTab.cs
@@ -1,11 +1,9 @@
-using System.Collections.Generic;
-using UnityEngine;
 
 namespace Unity.AI.Assistant.UI.Editor.Scripts.Components
 {
     class ProjectPopupTab : SearchableTab
     {
-        public override string Filter => "p: ";
+        public override SearchContextWrapper[] SearchProviders => new[] { ProjectSearchContext };
         public override string DefaultLabel => "Project";
         public override int Order => 1;