+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-pre.5] - 2025-05-14
+- Improved Agent deletion logic
+- Improved search window performance
+- Improved automatic conversation naming
+- Fixed issue with conversation history after editing agent actions
+- Other small bug fixes
+
 ## [1.0.0-pre.4] - 2025-05-13
 - Update AI-Toolkit to 1.0.0.-pre.7
 - Added console tab attachments